pyproject.toml
[tool.poetry]
name = "starter-fastapi"
version = "0.1.0"
description = "Starter template for FastAPI app"
authors = ["xbmlz <997909544@qq.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.82.0"
uvicorn = "^0.18.3"
python-dotenv = "^0.21.0"
alembic = "^1.8.1"
sqlalchemy = "^1.4.44"
pymysql = "^1.0.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.991"
black = {version = "^22.10.0", allow-prereleases = true}

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun.mirrors"
url = "https://mirrors.aliyun.com/pypi/simple"